refactor(core/embed): introduce display_utils.c
[no changelog]
