+++ core/SConscript.bootloader_ci
@@ -75,6 +75,7 @@ CPPPATH_MOD += [
 ]
 SOURCE_MOD += [
     'embed/lib/display.c',
+    'embed/lib/display_utils.c',
     'embed/lib/terminal.c',
     'embed/lib/colors.c',
     'embed/lib/fonts/fonts.c',
