+++ core/SConscript.prodtest
@@ -80,6 +80,7 @@ CPPPATH_MOD += [
 
 SOURCE_MOD += [
     'embed/lib/display.c',
+    'embed/lib/display_utils.c',
     'embed/lib/colors.c',
     'embed/lib/fonts/fonts.c',
     'embed/lib/fonts/font_bitmap.c',
