+++ core/embed/lib/display_utils.c
@@ -1,0 +1,31 @@
+/*
+ * This file is part of the Trezor project, https://trezor.io/
+ *
+ * Copyright (c) SatoshiLabs
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "common.h"
+#include "display.h"
+
+void display_fade(int start, int end, int delay) {
+#ifdef USE_BACKLIGHT
+  for (int i = 0; i < 100; i++) {
+    display_backlight(start + i * (end - start) / 100);
+    hal_delay(delay / 100);
+  }
+  display_backlight(end);
+#endif
+}
